Remove some tests that weren't helping
compiler-macroexpand-environment had a test for notinline, and
constant-form-value checked for macroexpansion of forms passed to constantp. The
former doesn't seem to be specified, i.e. whether (compiler-macro-function foo
env) where foo has been declared notinline returns NIL is unspecified;
futhermore some thought makes it seem ambiguous whether it even should.
constantp by contrast is explicitly allowed to not macroexpand forms passed to
it, so the test was dependent on unspecified behavior. No point in keeping that.

Other tests and test failures are still in place. I also switched the orders of
a few arguments because of the way 5am's magic IS works, while I was at it.